+++ src/core/tamper/caret.py
@@ -17,35 +17,35 @@ import sys
 from src.utils import settings
 
 """
-Adds single quotes (') between the characters of the generated payloads.
+Adds caret symbol (^) between the characters of the generated payloads.
 Notes:
-  * This tamper script works against *nix targets.
+  * This tamper script works against windows targets.
 """
 
-script_name = "singlequotes"
+script_name = "caret"
 
 if not settings.TAMPER_SCRIPTS[script_name]:
   settings.TAMPER_SCRIPTS[script_name] = True
 
 def transform(payload):
-  def add_single_quotes(payload):
+  def add_caret_symbol(payload):
     settings.TAMPER_SCRIPTS[script_name] = True
+    if re.compile("\w+").findall(payload):
+      if str(len(max(re.compile("\w+").findall(payload), key=lambda word: len(word)))) >= 5000:  
+        long_string = max(re.compile("\w+").findall(payload), key=lambda word: len(word))
+
     rep = {
-            "''i''f": "if", 
-            "''t''h''e''n": "then",
-            "''e''l''s''e": "else",
-            "''f''i": "fi",
-            "''s''t''r": "str",
-            "''c''m''d": "cmd",
-            "''c''ha''r": "char"
+            "^^": "^",
+            '"^t""^o""^k""^e""^n""^s"': '"t"^"o"^"k"^"e"^"n"^"s"',
+            re.sub(r'([b-zD-Z])', r'^\1', long_string) : long_string.replace("^","")
           }
-    payload = re.sub(r'([b-zD-Z])', r"''\1", payload)
+    payload = re.sub(r'([b-zD-Z])', r'^\1', payload)
     rep = dict((re.escape(k), v) for k, v in rep.iteritems())
     pattern = re.compile("|".join(rep.keys()))
     payload = pattern.sub(lambda m: rep[re.escape(m.group(0))], payload)
     return payload
 
-  if settings.TARGET_OS != "win":
+  if settings.TARGET_OS == "win":
     if settings.EVAL_BASED_STATE != False:
       if settings.TRANFROM_PAYLOAD == None:
         settings.TRANFROM_PAYLOAD = False
@@ -56,16 +56,15 @@ def transform(payload):
     else:
       settings.TRANFROM_PAYLOAD = True
       if settings.TRANFROM_PAYLOAD:
-        payload = add_single_quotes(payload)
-
+        payload = add_caret_symbol(payload)
   else:
     if settings.TRANFROM_PAYLOAD == None:
       settings.TRANFROM_PAYLOAD = False
-      warn_msg = "Windows target host(s), does not support the '"+ script_name  +".py' tamper script."
+      warn_msg = "*nix target host(s), does not support the '"+ script_name  +".py' tamper script."
       sys.stdout.write("\r" + settings.print_warning_msg(warn_msg))
       sys.stdout.flush() 
       print
 
   return payload
-  
+    
 # eof