+++ src/utils/logs.py
@@ -58,18 +58,26 @@ pass
 Save command history.
 """
 def save_cmd_history():
-  cli_history = os.path.expanduser(settings.CLI_HISTORY)
-  if os.path.exists(cli_history):
-    readline.write_history_file(cli_history)
+  try:
+    cli_history = os.path.expanduser(settings.CLI_HISTORY)
+    if os.path.exists(cli_history):
+      readline.write_history_file(cli_history)
+  except IOError, err_msg:
+    print settings.print_critical_msg(str(err_msg.args[0]).split("] ")[1] + ".")
+    sys.exit(0)
 
 """
 Load commands from history.
 """
 def load_cmd_history():
-  cli_history = os.path.expanduser(settings.CLI_HISTORY)
-  if os.path.exists(cli_history):
-    readline.read_history_file(cli_history)
-    
+  try:
+    cli_history = os.path.expanduser(settings.CLI_HISTORY)
+    if os.path.exists(cli_history):
+      readline.read_history_file(cli_history)
+  except IOError, err_msg:
+    print settings.print_critical_msg(str(err_msg.args[0]).split("] ")[1] + ".")
+    sys.exit(0)
+   
 """
 Create log files
 """
@@ -82,19 +90,26 @@ def create_log_file(url, output_dir):
     host = parts[1].split('/', 1)[0]
   except IndexError:
     host = parts[0].split('/', 1)[0]
-
+  except OSError, err_msg:
+    error_msg = str(err_msg.args[0]).split("] ")[1] + "."
+    print settings.print_critical_msg(error_msg)
+    raise SystemExit()
+    
   # Check if port is defined to host.
   if ":" in host:
     host = host.replace(":","_")
-
   try:
     os.stat(output_dir + host + "/")
   except:
     try:
       os.mkdir(output_dir + host + "/")
-    except OSError, err_msg:
-      print settings.print_critical_msg(str(err_msg.args[0]).split("] ")[1] + ".")
-      sys.exit(0)
+    except:
+      try:
+        error_msg = str(err_msg.args[0]).split("] ")[1] + "."
+      except:
+        error_msg = str(err_msg.args[0]) + "."
+      print settings.print_critical_msg(error_msg)
+      raise SystemExit()
 
   # Create cli history file if does not exist.
   settings.CLI_HISTORY = output_dir + host + "/" + "cli_history"